+++ samples/SemanticKernel/Encamina.Enmarcha.Samples.SemanticKernel.Text/Program.cs
@@ -1,4 +1,4 @@
-﻿using Encamina.Enmarcha.SemanticKernel.Abstractions;
+﻿using Encamina.Enmarcha.AI.OpenAI.Azure;
 using Encamina.Enmarcha.SemanticKernel.Plugins.Text;
 
 using Microsoft.Extensions.Configuration;
@@ -11,7 +11,7 @@ namespace Encamina.Enmarcha.Samples.SemanticKernel.Text;
 
 internal static class Program
 {
-    private static async Task Main(string[] args)
+    private static async Task Main(string[] _)
     {
         // Create and configure builder
         var hostBuilder = new HostBuilder()
@@ -26,8 +26,8 @@ internal static class Program
             services.AddScoped(_ =>
             {
                 // Get semantic kernel options
-                var options = hostContext.Configuration.GetRequiredSection(nameof(SemanticKernelOptions)).Get<SemanticKernelOptions>()
-                ?? throw new InvalidOperationException(@$"Missing configuration for {nameof(SemanticKernelOptions)}");
+                var options = hostContext.Configuration.GetRequiredSection(nameof(AzureOpenAIOptions)).Get<AzureOpenAIOptions>()
+                ?? throw new InvalidOperationException(@$"Missing configuration for {nameof(AzureOpenAIOptions)}");
 
                 // Initialize semantic kernel
                 var kernel = Kernel.CreateBuilder()