+++ src/Encamina.Enmarcha.SemanticKernel.Plugins.Chat/KernelExtensions.cs
@@ -1,5 +1,6 @@
 ﻿using CommunityToolkit.Diagnostics;
 
+using Encamina.Enmarcha.AI.OpenAI.Abstractions;
 using Encamina.Enmarcha.Data.Abstractions;
 using Encamina.Enmarcha.Data.Cosmos;
 
@@ -25,13 +26,6 @@ public static class KernelExtensions
     /// This extension method uses a «Service Location» pattern provided by the <see cref="IServiceProvider"/> to resolve the following dependencies:
     /// <list type="bullet">
     ///     <item>
-    ///         <term>SemanticKernelOptions</term>
-    ///         <description>
-    ///             A required dependency of type <see cref="SemanticKernelOptions"/> used to retrieve the configurations for the <see cref="Kernel"/>. This dependency
-    ///             should be added using any of the <see cref="OptionsServiceCollectionExtensions.AddOptions"/> extension method.
-    ///         </description>
-    ///     </item>
-    ///     <item>
     ///         <term>ChatWithHistoryPluginOptions</term>
     ///         <description>
     ///             A required dependency of type <see cref="ChatWithHistoryPluginOptions"/> used to retrieve the configuration options for this plugin. This dependency should
@@ -49,22 +43,22 @@ public static class KernelExtensions
     /// </remarks>
     /// <param name="kernel">The <see cref="Kernel"/> instance to add this plugin.</param>
     /// <param name="serviceProvider">A <see cref="IServiceProvider"/> to resolve the dependencies.</param>
+    /// <param name="openAIOptions">Configuration options for OpenAI services.</param>
     /// <param name="cosmosContainer">The name of the Cosmos DB container to store the chat history messages.</param>
     /// <param name="tokensLengthFunction">
     /// A function to calculate the length by tokens of the chat messages. These functions are usually available in the «mixin» interface <see cref="ILengthFunctions"/>.
     /// </param>
     /// <returns>A list of all the functions found in this plugin, indexed by function name.</returns>
-    public static KernelPlugin ImportChatWithHistoryPluginUsingCosmosDb(this Kernel kernel, IServiceProvider serviceProvider, string cosmosContainer, Func<string, int> tokensLengthFunction)
+    public static KernelPlugin ImportChatWithHistoryPluginUsingCosmosDb(this Kernel kernel, IServiceProvider serviceProvider, OpenAIOptions openAIOptions, string cosmosContainer, Func<string, int> tokensLengthFunction)
     {
         Guard.IsNotNull(serviceProvider);
         Guard.IsNotNull(tokensLengthFunction);
         Guard.IsNotNullOrWhiteSpace(cosmosContainer);
 
-        var semanticKernelOptions = serviceProvider.GetRequiredService<IOptionsMonitor<SemanticKernelOptions>>().CurrentValue;
         var chatWithHistoryPluginOptions = serviceProvider.GetRequiredService<IOptionsMonitor<ChatWithHistoryPluginOptions>>();
         var chatMessagesHistoryRepository = serviceProvider.GetRequiredService<ICosmosRepositoryFactory>().Create<ChatMessageHistoryRecord>(cosmosContainer);
 
-        var chatWithHistoryPlugin = new ChatWithHistoryPlugin(kernel, semanticKernelOptions.ChatModelName, tokensLengthFunction, chatMessagesHistoryRepository, chatWithHistoryPluginOptions);
+        var chatWithHistoryPlugin = new ChatWithHistoryPlugin(kernel, openAIOptions.ChatModelName, tokensLengthFunction, chatMessagesHistoryRepository, chatWithHistoryPluginOptions);
 
         return kernel.ImportPluginFromObject(chatWithHistoryPlugin, PluginsInfo.ChatWithHistoryPlugin.Name);
     }